chore(*): extract spring data configuration
