+++ src/main/java/sleppynavigators/studyupbackend/infrastructure/common/AuditingConfig.java
@@ -1,13 +1,15 @@
-package sleppynavigators.studyupbackend.infrastructure.common.config;
+package sleppynavigators.studyupbackend.infrastructure.common;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.mongodb.config.EnableMongoAuditing;
 
 @Configuration
 @EnableJpaAuditing
-public class JpaConfig {
+@EnableMongoAuditing
+public class AuditingConfig {
 
     @Bean
     public AuditorAware<?> auditorProvider() {
